- PAL'ify the GetCommandLineParametersAsString() to handle windows specific behavior - Minor optimization for argument string vector initialization
