+++ Code/Framework/AzFramework/Platform/Android/AzFramework/Process/ProcessWatcher_Android.cpp
@@ -8,7 +8,7 @@
 
 #include <AzFramework/Process/ProcessWatcher.h>
 #include <AzFramework/Process/ProcessCommunicator.h>
-
+#include <AzFramework/StringFunc/StringFunc.h>
 
 namespace AzFramework
 {
@@ -82,5 +82,24 @@ namespace AzFramework
     void ProcessWatcher::TerminateProcess([[maybe_unused]] AZ::u32 exitCode)
     {
 
+    }
+
+    AZStd::string ProcessLauncher::ProcessLaunchInfo::GetCommandLineParametersAsString() const
+    {
+        struct CommandLineParametersVisitor
+        {
+            AZStd::string operator()(const AZStd::string& commandLine) const
+            {
+                return commandLine;
+            }
+
+            AZStd::string operator()(const AZStd::vector<AZStd::string>& commandLineArray) const
+            {
+                AZStd::string commandLineResult;
+                AzFramework::StringFunc::Join(commandLineResult, commandLineArray.begin(), commandLineArray.end(), " ");
+                return commandLineResult;
+            }
+        };
+        return AZStd::visit(CommandLineParametersVisitor{}, m_commandlineParameters);
     }
 } //namespace AzFramework
