+++ Code/Framework/AzFramework/Platform/Windows/AzFramework/Process/ProcessWatcher_Win.cpp
@@ -14,6 +14,7 @@
 #include <AzFramework/Process/ProcessWatcher.h>
 #include <AzFramework/Process/ProcessCommunicator.h>
 #include <AzFramework/Process/ProcessCommon.h>
+#include <AzFramework/StringFunc/StringFunc.h>
 
 #include <iostream>
 
@@ -353,6 +354,50 @@ namespace AzFramework
         else
         {
             ::TerminateProcess(m_pWatcherData->processInformation.hProcess, exitCode);
+        }
+    }
+
+        namespace ProcessLauncher
+    {
+        AZStd::string ProcessLaunchInfo::GetCommandLineParametersAsString() const
+        {
+            struct CommandLineParametersVisitor
+            {
+                AZStd::string operator()(const AZStd::string& commandLine) const
+                {
+                    return commandLine;
+                }
+
+                AZStd::string operator()(const AZStd::vector<AZStd::string>& commandLineArray) const
+                {
+                    AZStd::string commandLineResult;
+
+                    // When re-constructing a command line from an argument list (on windows), if an argument
+                    // is double-quoted, then the double-quotes must be escaped properly otherwise
+                    // it will be absorbed by the native argument parser and possibly evaluated as
+                    // multiple values for arguments
+                    AZStd::string_view escapedDoubleQuote = R"("\")";
+
+                    AZStd::vector<AZStd::string> preprocessed_command_array;
+                    for (const auto& commandArg : commandLineArray)
+                    {
+                        AZStd::string replacedArg = commandArg;
+                        auto replacePos = replacedArg.find_first_of('"');
+                        while (replacePos != replacedArg.npos)
+                        {
+                            replacedArg.replace(replacePos, 1, escapedDoubleQuote.data(), escapedDoubleQuote.size());
+                            replacePos += escapedDoubleQuote.size();
+                            replacePos = replacedArg.find('"', replacePos);
+                        }
+                        preprocessed_command_array.emplace_back(replacedArg);
+                    }
+                    AzFramework::StringFunc::Join(commandLineResult, preprocessed_command_array.begin(), preprocessed_command_array.end(), " ");
+
+                    return commandLineResult;
+                }
+            };
+
+            return AZStd::visit(CommandLineParametersVisitor{}, m_commandlineParameters);
         }
     }
 } // namespace AzFramework
